requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyodps>=0.11.0
joblib>=1.1.0
pyyaml>=6.0
requests>=2.25.0
jupyter>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0